# back/docs/swagger.yaml
definitions:
  admin.CreateTeamInput:
    properties:
      Name:
        type: string
    required:
    - Name
    type: object
  admin.CreateUserInput:
    properties:
      Email:
        type: string
      FirstName:
        type: string
      LastName:
        type: string
      Password:
        type: string
      Role:
        type: string
      Team:
        type: string
    required:
    - Email
    - FirstName
    - LastName
    - Password
    - Role
    type: object
  admin.UpdateTeamInput:
    properties:
      CurrentName:
        type: string
      NewName:
        type: string
    required:
    - CurrentName
    - NewName
    type: object
  admin.UpdateUserTeamInput:
    properties:
      TeamId:
        type: string
    required:
    - TeamId
    type: object
  authentification.AuthentificationUserInput:
    properties:
      Email:
        type: string
      Password:
        type: string
    required:
    - Email
    - Password
    type: object
  authentification.AuthentificationUserOutput:
    properties:
      _id:
        type: string
      Email:
        type: string
      FirstName:
        type: string
      LastName:
        type: string
      Role:
        type: string
      Team:
        type: string
      Token:
        type: string
    required:
    - Email
    - FirstName
    - LastName
    - Role
    - Token
    - _id
    type: object
  presence.CreatePresenceInput:
    properties:
      Timestamp:
        type: string
      Type:
        type: string
    required:
    - Timestamp
    - Type
    type: object
  presence.PresenceOutput:
    properties:
      Timestamp:
        type: string
      Type:
        type: string
    required:
    - Timestamp
    - Type
    type: object
  teams.TeamOutput:
    properties:
      _id:
        type: string
      Name:
        type: string
    required:
    - Name
    - _id
    type: object
  users.UpdateUserInput:
    properties:
      FirstName:
        type: string
      LastName:
        type: string
      Password:
        type: string
    type: object
  users.UserOutput:
    properties:
      _id:
        type: string
      Email:
        type: string
      FirstName:
        type: string
      LastName:
        type: string
      Role:
        type: string
      Team:
        type: string
    required:
    - Email
    - FirstName
    - LastName
    - Role
    - _id
    type: object
host: localhost:8080
info:
  contact: {}
  title: Time-Manager API
  version: "0.1"
paths:
  /api/admin/{id}:
    get:
      description: Fetch a single user and returns its info if successful
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserOutput'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Fetch a single user by id
      tags:
      - Admin
  /api/admin/all/team:
    get:
      description: Returns a list of all teams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/teams.TeamOutput'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all teams
      tags:
      - Admin
  /api/admin/all/user:
    get:
      description: Returns a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserOutput'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Admin
  /api/admin/team/create:
    post:
      consumes:
      - application/json
      description: Create a single new team from a specific team Input. Will return
        an error if team with same name already exists.
      parameters:
      - description: Team to create
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/admin.CreateTeamInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teams.TeamOutput'
        "400":
          description: Invalid input
          schema:
            type: string
        "409":
          description: 'Conflict : team already exists'
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a single new team
      tags:
      - Admin
  /api/admin/team/update:
    put:
      consumes:
      - application/json
      description: Update a single existing team from a specific team Input. Will
        return corresponding team if no modification is done
      parameters:
      - description: Current team name and New team name
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateTeamInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teams.TeamOutput'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "409":
          description: 'Conflict : team already exists'
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a single existing team
      tags:
      - Admin
  /api/admin/team/users/{name}:
    get:
      description: Return all users from the specified team. Will return an error
        if team doesn't exist. Will return an empty array if no user is in the team
      parameters:
      - description: Team name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserOutput'
            type: array
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Team not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Return all users from a team
      tags:
      - Admin
  /api/admin/update/team/user/{id}:
    put:
      consumes:
      - application/json
      description: Update the team of a user identified by id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      - description: Team to associate with the user
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateUserTeamInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserOutput'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a user's team
      tags:
      - Admin
  /api/admin/user/create:
    post:
      consumes:
      - application/json
      description: Create a single new user from a specific user Input. Will return
        an error if user already exists.
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/admin.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserOutput'
        "400":
          description: Invalid input
          schema:
            type: string
        "409":
          description: 'Conflict : user already exists'
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a single new user
      tags:
      - Admin
  /api/authentification:
    post:
      consumes:
      - application/json
      description: Authenticates a user using email and password, and returns a JWT
        token if successful
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/authentification.AuthentificationUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authentification.AuthentificationUserOutput'
        "400":
          description: Invalid input
        "404":
          description: User not found or wrong credentials
        "500":
          description: Internal server error
      summary: Authenticate a user
      tags:
      - Authentication
  /api/manager/team/user/presence/{id}:
    get:
      description: Returns all presence records for the specified user if the user
        belongs to the manager's team
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presence.PresenceOutput'
            type: array
        "400":
          description: Invalid input / not member of team
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get presences for a user
      tags:
      - Manager
  /api/manager/team/users/{name}:
    get:
      description: Return all users belonging to the team the manager manages (team
        is derived from the manager's token/claim)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserOutput'
            type: array
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Return all users from the manager's team
      tags:
      - Manager
  /api/presence:
    get:
      description: Returns all presences belonging to the user identified by the JWT
        token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presence.PresenceOutput'
            type: array
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Return all presences for the authenticated user
      tags:
      - Presence
  /api/presence/create:
    post:
      consumes:
      - application/json
      description: Create a new presence entry for the authenticated user. Returns
        conflict if duplicate presence exists.
      parameters:
      - description: Presence to create
        in: body
        name: presence
        required: true
        schema:
          $ref: '#/definitions/presence.CreatePresenceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presence.PresenceOutput'
        "400":
          description: Invalid input
          schema:
            type: string
        "409":
          description: 'Conflict : presence with same information already exists'
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new presence record
      tags:
      - Presence
  /api/team/name:
    get:
      description: Fetch a single team and returns its info if successful
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/teams.TeamOutput'
        "400":
          description: Invalid Input
        "404":
          description: Team not found
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Fetch a single team
      tags:
      - Team
  /api/user:
    get:
      description: Fetch a single user and returns its info if successful
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserOutput'
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Fetch a single user
      tags:
      - User
  /api/user/update:
    post:
      consumes:
      - application/json
      description: Update a single existing user from a specific user Input. Will
        return corresponding user if no modification is done
      parameters:
      - description: User fields to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserOutput'
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a single existing user
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: api_token
    type: apiKey
swagger: "2.0"